feat(Serenytics): reformating code
