+++ v0/destinations/serenytics/data/SerenyticsAliasConfig.json
@@ -23,7 +23,6 @@
   {
     "destKey": "original_timestamp",
     "sourceKeys": "originalTimestamp",
-    "sourceFromGenericMap": true,
     "required": false
   },
   {
