+++ v0/destinations/serenytics/data/SerenyticsIdentifyConfig.json
@@ -52,7 +52,6 @@
   {
     "destKey": "original_timestamp",
     "sourceKeys": "originalTimestamp",
-    "sourceFromGenericMap": true,
     "required": false
   },
   {
